#169 minor refactorings
